refactored base_forecast and prophet_forecast to enable easier testing
